Significant revision of abd2latex.py
Significantly revised abd2latex.py based on code review with Kevin.
(1) Defined two functions, get_elem_name() and reformat_files(), to handle converting the element's numerical ID to a text string, and to process and reformat the input files, respectively.
(2) Imported argparse to process the arguments and provide usage help for the user.
(3) To make it more unix-like, the output is now by default printed to stdout, unless an output file is specified with the -o option.
(4) Used the "with ___ as ___:" context to write the reformatted data to the output file. This method could be very memory inefficient for large files, but the text files being reformatted are never going to be that large.
(5) Attempted to make the variable names more intuitive and straightforward.